# src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "LuisGonzales",
        "version": "1.0.0"
    },
    "paths": {
        "/locations": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {

                                    }
                                },
                                "examples": {
                                    "all": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "name": "International Inc Corporate Office",
                                                "location": {
                                                    "lat": 51.5013673,
                                                    "lng": -0.1440787
                                                },
                                                "type": "headquarter",
                                                "status": "1"
                                            },
                                            {
                                                "id": 2,
                                                "name": "International Inc North America",
                                                "location": {
                                                    "lat": 40.6976701,
                                                    "lng": -74.259876
                                                },
                                                "type": "office",
                                                "status": "1"
                                            },
                                            {
                                                "id": 3,
                                                "name": "International Inc France",
                                                "location": {
                                                    "lat": 48.859,
                                                    "lng": 2.2069746
                                                },
                                                "type": "office",
                                                "status": "1"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "todas las ubicaciones\n"
                    }
                }
            }
        },
        "/locations/{id}": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/location"
                                }
                            }
                        },
                        "description": "1"
                    }
                }
            },
            "parameters": [
                {
                    "name": "id",
                    "schema": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/locations/phone/{id}/": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/location"
                                }
                            }
                        },
                        "description": "2"
                    }
                }
            },
            "parameters": [
                {
                    "name": "id",
                    "schema": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        }
    },
    "components": {
        "schemas": {
            "location": {
                "title": "Root Type for location",
                "description": "",
                "type": "object",
                "properties": {
                    "id": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "location": {
                        "type": "object",
                        "properties": {
                            "lat": {
                                "format": "double",
                                "type": "number"
                            },
                            "lng": {
                                "format": "double",
                                "type": "number"
                            }
                        }
                    },
                    "type": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "example": {
                    "id": 1,
                    "name": "International Inc Corporate Office",
                    "location": {
                        "lat": 51.5013673,
                        "lng": -0.1440787
                    },
                    "type": "headquarter",
                    "status": "1"
                }
            }
        }
    }
}